requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==0.7.16
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.8
astropy==7.0.1
astropy-iers-data==0.2025.2.17.0.34.13
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==23.2.0
autopep8==2.3.2
babel==2.17.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
beautifulsoup4==4.13.3
billiard==4.2.0
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blessed==1.20.0
blinker==1.9.0
blurb==2.0.0
bokeh==3.6.3
boltons==25.0.0
Bottleneck==1.4.2
brotlipy==0.7.0
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.2
cargo==0.3
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
commonmark==0.9.1
con==1.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
docx==0.2.4
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
etils==1.12.0
fastapi-cli==0.0.7
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
flax==0.10.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
google_search_results==2.4.2
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.0
humanize==4.12.1
id==1.5.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jinxed==1.3.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.9.1
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.3.19
langchain-core==0.3.36
langchain-text-splitters==0.3.6
langsmith==0.3.8
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==2.2.3
openai==1.63.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
pathspec==0.12.1
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pprintpp==0.4.0
ppt==0.0.7.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyerfa==2.0.1.5
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyJWT==2.10.1
PyMuPDF==1.25.5
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.4.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doc==2020.10.1
python-docs-theme==2024.12
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rich-toolkit==0.14.1
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sendgrid==6.11.0
serpapi==0.1.5
setuptools==75.8.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==2.4.4
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
starkbank-ecdsa==2.2.0
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-intel==0.0.1
tensorstore==0.1.71
termcolor==3.0.0
textblob==0.19.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
treescope==0.1.9
twine==6.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-pytz==2025.2.0.20250326
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.30.0
visdom==0.2.4
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
XlsxWriter==3.2.2
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
